#criando regras de negocio

+++ src/main/java/com/example/demo/applications/CategoriasApplication.java
@@ -1,14 +1,17 @@
 package com.example.demo.applications;
 
+import com.example.demo.config.RegraNegocioException;
 import com.example.demo.entities.Categoria;
 import com.example.demo.entities.Desafio;
 import com.example.demo.interfaces.ICategorias;
 import com.example.demo.repositories.CategoriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class CategoriasApplication implements ICategorias {
@@ -23,12 +26,22 @@ public class CategoriasApplication implements ICategorias {
 
     @Override
     public Categoria salvar(Categoria categoria) {
+        String nome = categoria.getNome();
+        if (nome == null || nome.trim().isEmpty() || nome.length() < 3) {
+            throw new RegraNegocioException("Nome da categoria inválido.");
+        }
+        if (existePorNome(categoria.getNome())) {
+            throw new RegraNegocioException("Categoria com este nome já existe.");
+        }
         return categoriaRepository.save(categoria);
     }
 
     @Override
     public Categoria buscarPorId(int id) {
-        return categoriaRepository.findById(id).orElseThrow();
+        return categoriaRepository.findById(id).orElseThrow(()->new RegraNegocioException(
+                "Categoria não encontrada com ID: " + id
+                )
+        );
     }
 
     @Override
@@ -38,11 +51,18 @@ public class CategoriasApplication implements ICategorias {
 
     @Override
     public List<Categoria> listarTodos() {
-        return categoriaRepository.findAll();
+        return categoriaRepository.findAll(Sort.by("nome").ascending());
     }
 
     @Override
     public void deletar(int id) {
+        Categoria categoria = buscarPorId(id);
+        List<Desafio> desafios = listarDesafiosPorCategoriaId(id);
+
+        boolean temDesafiosAtivos = desafios.stream().anyMatch(d -> d.getStatus().equals("ATIVO"));
+        if (temDesafiosAtivos) {
+            throw new RegraNegocioException("Não é possível excluir categoria com desafios ativos.");
+        }
         categoriaRepository.deleteById(id);
     }
 
@@ -65,4 +85,11 @@ public class CategoriasApplication implements ICategorias {
     public List<Desafio> listarDesafiosPorCategoriaId(int idCategoria) {
         return categoriaRepository.findDesafiosPorCategoriaId(idCategoria);
     }
+
+    public List<Desafio> listarDesafiosAtivosPorCategoria(int idCategoria) {
+        return listarDesafiosPorCategoriaId(idCategoria).stream()
+                .filter(d -> d.getStatus().equals("ATIVO"))
+                .collect(Collectors.toList());
+    }
+
 }
